requirements.txt
asgiref==3.2.3
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
confusable-homoglyphs==3.2.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.41.0
django-bootstrap4==1.1.1
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
gunicorn==20.0.4
idna==2.8
oauthlib==3.1.0
Pillow==6.2.1
pkg-resources==0.0.0
psycopg2==2.8.4
python-decouple==3.3
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
whitenoise==5.0.1
